requirements.txt
django
untangle
pytz
requests
transliterate
django-ipware